// src/app/modules/configuration/finance/models/manage-fee-configuration-view.ts
/**
 * Fee
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { SchoolFeeConfigurationView } from './school-fee-configuration-view';


/**
 * ManageFeeConfigurationView
 */
export interface ManageFeeConfigurationView { 
    /**
     * Gets or Sets ClassId
     */
    classId?: number;
    /**
     * Gets or Sets SchoolBoardId
     */
    schoolBoardId?: number;
    /**
     * Gets or Sets SchoolFeeConfiguration
     */
    schoolFeeConfiguration?: Array<SchoolFeeConfigurationView>;
}