// src/app/modules/configuration/finance/models/clone-school-fee-view.ts
/**
 * Fee
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


/**
 * CloneSchoolFeeView
 */
export interface CloneSchoolFeeView { 
    /**
     * Gets or Sets SourceClassId
     */
    sourceClassId?: number;
    /**
     * Gets or Sets sourceSchoolBoardId
     */
    sourceSchoolBoardId?: number;
    /**
     * Gets or Sets destinationClassIds
     */
    destinationClassIds?: Array<number>;
      /**
     * Gets or Sets destinationSchoolBoardId
     */
    destinationSchoolBoardId?: number;
}