// src/app/modules/configuration/finance/models/school-fee-view-model.ts
import { SchoolFeeAssignmentViewModel } from './school-fee-assignment-view-model';

/**
 * Fee
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


/**
 * SchoolFeeViewModel
 */
export interface SchoolFeeViewModel { 
    /**
     * Gets or Sets ClassId
     */
    classId?: number;
    /**
     * Gets or Sets BoardId
     */
    schoolBoardId?: number;
    /**
     * Gets or Sets SchoolFeeId
     */
    schoolFeeId?: string;
    /**
     * Gets or Sets FeeTypeId
     */
    feeTypeId?: string;
    /**
     * Gets or Sets Name
     */
    name?: string;
    /**
     * Gets or Sets TotalAmount
     */
    feeAmount?: number;
    /**
     * Gets or Sets SchoolFeeAssignments
     */
    schoolFeeAssignments?: Array<SchoolFeeAssignmentViewModel>;
}